+++ src/lib/data/user-dto.ts
@@ -1,7 +1,8 @@
 import "server-only";
-
+import { auth, Role } from "@/auth";
 import { User } from "@/generated/prisma";
 import prisma from "../prisma";
+import { maskName } from "../util";
 
 export type UserDTO = {
   id: string;
@@ -24,6 +25,12 @@ export function prismaUserToDTO(user: User): UserDTO {
 }
 
 export async function getUserDTO(id: string): Promise<UserDTO | null> {
+  const session = await auth();
+
+  if (!session || (session.user.id !== id && session.user.role !== Role.ADMIN)) {
+    return null;
+  }
+
   const user = await prisma.user.findUnique({
     where: { id },
   });
@@ -36,23 +43,41 @@ export async function getUserDTO(id: string): Promise<UserDTO | null> {
 }
 
 export async function getAllUserDTOs(): Promise<UserDTO[]> {
+  const session = await auth();
+
+  if (!session || session.user.role !== Role.ADMIN) {
+    return [];
+  }
+
   const users = await prisma.user.findMany({});
 
   return users.map(prismaUserToDTO);
 }
 
 export async function getAllUserNames(): Promise<{ id: string; name: string }[]> {
+  const session = await auth();
+
   const users = await prisma.user.findMany({
     select: { id: true, name: true },
   });
 
-  return users.map(user => ({ id: user.id, name: user.name }));
+  return users.map(user => ({ id: user.id, name: !session ? maskName(user.name) : user.name }));
 }
 
 export async function createUser(data: {
   email: string;
   name: string;
 }): Promise<UserDTO> {
+  const session = await auth();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
+
+  if (session.user.role !== Role.ADMIN) {
+    throw new Error("Forbidden");
+  }
+
   const user = await prisma.user.create({
     data: {
       email: data.email,
@@ -67,6 +92,16 @@ export async function updateUser(id: string, data: {
   email: string;
   name: string;
 }): Promise<UserDTO | null> {
+  const session = await auth();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
+
+  if (session.user.role !== Role.ADMIN) {
+    throw new Error("Forbidden");
+  }
+
   const user = await prisma.user.update({
     where: { id },
     data: {
@@ -83,6 +118,16 @@ export async function updateUser(id: string, data: {
 }
 
 export async function deleteUser(id: string): Promise<UserDTO> {
+  const session = await auth();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
+
+  if (session.user.role !== Role.ADMIN) {
+    throw new Error("Forbidden");
+  }
+
   const user = await prisma.user.delete({
     where: { id },
   });
@@ -91,6 +136,16 @@ export async function deleteUser(id: string): Promise<UserDTO> {
 }
 
 export async function deleteUsers(ids: string[]) {
+  const session = await auth();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
+
+  if (session.user.role !== Role.ADMIN) {
+    throw new Error("Forbidden");
+  }
+
   const payload = await prisma.user.deleteMany({
     where: { id: { in: ids } },
   });
